eBook conversion - Fix handling of unknown <div> types
The handling of <div> tags which contain a type attribute which is not
one of those explicitly handled has been fixed to prevent an exception.
osisID attribure values containing two discrete verse references are now
handled.  Changes to footnote handling to cater for the the tags
generated by usfm2osis.py have been started but are not currently
working correctly,