Ebook conversion book title extraction fixes
Biblical book titles can now be extracted from the OSIS file even when
located after the initial <chapter> and <verse> tags. There has also
been a fix to correct the generation of the TOC entry for an extracted
book title containing a line break for epub and mobi (fix is not
possible for fb2 due to Calibre limitations). A problem which could
result in 2 TOC entries being generated for a chapter has also been
fixed.