eBook conversion - fixed problems with chapter numbers and book intros
Fix for a problem where chapter/Psalm numbers were not being generated
where a chapter/Psalm starts with a poetic line. The layout of initial
poetic lines next to chapter/Psalm numbers has been improved. There is
also a fix for a problem which could result in book introductions not
being written.