eBook conversion: Fixes for testament intro problems
Fixes for problems associated with testament introductions outside
books, the most serious of which was failure to include the introduction
in the output. Also fixed a problem where specifying a testament title
in convert.txt resulted in an incorrectly generated table of contents if
there was as space between "Group<n>" and the following "=". Fixed a
further problem where a crash was caused if the last book group in the
OSIS file was empty.