eBooks - Fix misplaced book intro footnote
This fixes a problem which resulted in a footnote in the introduction of
the first book of a testament appearing at the end of the preceding
testament introduction with the footnote reference linking to the wrong
footnote. The following changes are also implemented:
a. 'TitlePage' divs can now been included within a Bible or testament
introduction.
b. Within a Bible or testament introduction, a table of contents entry
is now created for each div, unless "IntroInContents" is set to False.