eBook Conversion - Correct manifest id generation for images
This change corrects an error in the generation of manifest ids for
images in the OPF file. The final character of the id was being lost,
which could result in ids being duplicated. This was not causing a
problem, as Calibre corrected the error, but the change eliminates error
messages about duplicate ids.