eBooks - Create toc entries for individual glossary entries
Entries in the table of contents can now be created for individual
glossary entries. By default, these will be created for epub and mobi
output formats. To create these for FB2 output format also, add
"GlossEntriesInToc=True" in convert.txt - this will result in the layout
of the glossary entries being altered for FB2. To prevent the table of
contents entries being created for any output format, include
"GlossEntriesInToc=False" in convert.txt.