eBook changes relating to introduction
The following changes have been implemented:

1. Additional <div> types are now recognised as being valid for outer
<div>'s in Bible or testament introduction.

2. Previously, if a title matching the book title was found in the
introductiion of the first book of a book group, the text before this
was regarded as being the testament introduction. This behaviour is now
disabled by default, and is only enabled if convert.txt includes
"CombinedIntros=True".

3. When "CombinedIntros=True" is specified, there will now be a table of
contents entry generated for a testament introduction extracted from a
book introduction, provided that there is a title from which an entry
can be generated.

4. A title occurring in an introduction which has type="x-chapterLabel"
is now handled correctly. Previously, any such title would not appear in
the eBook.