eBook conversion - Fix problems associated with combined title/keyword
This is a fix for problems associated with the situation where, within a
glossary, when a glossary keyword is also a glossary title (i.e. '<seg
type="keyword">' is within a <title> tag. Previously this caused invalid
html to be generated, and the title did not appear at the correct level
in the Table of Contents. These problems are now fixed.